refactor: Move business logic from view_models to models/app_state
- Create models/app_state directory with proper module structure
- Move all business logic modules from view_models to app_state:
  - buffer_operations, cursor_manager, display_manager
  - ex_command_manager, http_manager, mode_manager
  - pane_manager, rendering_coordinator, settings_manager
- AppState now follows same pattern as PaneState with encapsulated logic
- ViewModel fields made pub(crate) for access from models/app_state
- Remove redundant yank_operations (already in buffer_operations)
- Update all imports and module references

This properly separates concerns - data and operations on that data
are together in the models layer, preventing feature envy and following
OOP principles. AppViewModel will become the coordination layer when
it absorbs AppController logic.
