+++ src/repl/models/app_state/core.rs
