+++ src/repl/models/app_state/display_manager.rs
